To make a sensor, universal plug-n-play, i.e. To make an interface with the help of an embedded system and a computer which can read any sensor when connected to a Serial RS232 port and display the type, model, manufacturer and the output data given by the sensor in a common format.
This process would require detecting the active COM ports, matching the baud rates of the sensor and the interface, then analyzing the raw output data acquired with the help of known formats specified for each different sensor and then finally displaying the output from the sensor in the desired format. Hence, broadly we need to accomplish the following- 4.1 Sensor Detection 4.2 Data Acquisition 4.3 Data Analysis 4.4 Data Storage Here we shall primarily concentrate on the Sensor Detection part of this project.